project/broadcaster/requirements.txt
asgiref==3.3.4
asyncio-nats-client==0.11.4
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
toml==0.10.2
Werkzeug==2.0.1
requests
